#161 Moved code from gtfs_stats to 2 new files
