+++ gtfs/gtfs_utils/gtfs_utils/gtfs_stats_conf.py
@@ -37,3 +37,12 @@ LOG_FOLDER = join(BASE_FOLDER, 'logs')
 
 PROFILE = False
 PROFILE_PATH = join(BASE_FOLDER, 'gtfs_stats_180601.prof')
+
+ALL_FOLDERS = [
+    DATA_FOLDER,
+    GTFS_FEEDS_PATH,
+    OUTPUT_DIR,
+    FILTERED_FEEDS_PATH,
+    LOG_FOLDER,
+    ARCHIVE_FOLDER
+]
